motor-backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
aiosqlite>=0.20.0
pydantic>=2.0.0
pyserial>=3.5
python-multipart>=0.0.9
pandas
scikit-learn
joblib
numpy
